This project is basically for those people who are not able to hear since sign languages are used most, so this projects help in reognizing differenent hand emojis with the differrent images . The model is trained in such a way that it chosses right hand emoji image with the hand you placed in the camera
This project is built on understanding the real terms communication barriers. Since emojis are all around used in the communication and have effective meaning in today’s world so this project is an ideologic approach to that idea. This is used in the sign language translation, since the message could be send through emojis, different signs can easily be send with this model.
I have worked really hard for the project and used python in the project. After completing the project, I have gathered confidence to work in any project related to OpenCV python and keras and TensorFlow.
I gathered the knowledge of convolution neural network in it which is a very important part of deep learning and since deep learning is a subfield of Artificial Intelligence.